Rename route to match existing conventions
In the main, when we are interacting with a 'thing' we do it on a route that matches. Viewing a bill run or return version? Then its `/bill-runs` and `/return-versions`.

In the case of when we come to view a return log, it will be `/return-logs`.

When we come to setup a 'thing', we put it on a `/setup` path _under_ the main path, for example, `/bill-runs/setup`.

So, we rename the route that was setup for return logs and the routing to match.

BTW we rename to setup because this feature, though called `edit a return log` isn't actually doing that. The first time we complete this journey, we will actually be setting up the first return submission for the return log. Any subsequent edits are just 'setting up' new versions of the return submission.

So, we change to `/setup`, not only to match existing conventions but also to give us some linguistic flexibility as we build out this feature.
