+++ app/controllers/return-logs-setup.controller.js
@@ -1,19 +1,17 @@
 'use strict'
 
 /**
- * Controller for /return-logs-edit endpoints
- * @module ReturnLogsEditController
+ * Controller for /return-logs/setup endpoints
+ * @module ReturnLogsSetupController
  */
 
 const InitiateSessionService = require('../services/return-logs-edit/initiate-session.service.js')
-
-const basePath = 'return-log-edit'
 
 async function setup(request, h) {
   const { returnLogId } = request.query
   const session = await InitiateSessionService.go(returnLogId)
 
-  return h.redirect(`/system/${basePath}/${session.id}/how-to-edit`)
+  return h.redirect(`/system/return-logs/setup/${session.id}/how-to-edit`)
 }
 
 module.exports = {
