+++ app/plugins/router.plugin.js
@@ -27,7 +27,7 @@ const NotificationsSetupRoutes = require('../routes/notifications-setup.routes.j
 const JobRoutes = require('../routes/jobs.routes.js')
 const LicenceRoutes = require('../routes/licence.routes.js')
 const MonitoringStationRoutes = require('../routes/monitoring-station.routes.js')
-const ReturnLogEditRoutes = require('../routes/return-logs-edit.routes.js')
+const ReturnLogSetupRoutes = require('../routes/return-logs-setup.routes.js')
 const ReturnVersionsSetupRoutes = require('../routes/return-versions-setup.routes.js')
 const ReturnVersionsRoutes = require('../routes/return-versions.routes.js')
 const RootRoutes = require('../routes/root.routes.js')
@@ -48,7 +48,7 @@ const routes = [
   ...LicenceRoutes,
   ...JobRoutes,
   ...MonitoringStationRoutes,
-  ...ReturnLogEditRoutes,
+  ...ReturnLogSetupRoutes,
   ...ReturnVersionsRoutes,
   ...ReturnVersionsSetupRoutes,
   ...DataRoutes,
