--- app/routes/return-logs-edit.routes.js
@@ -1,22 +1,0 @@
-'use strict'
-
-const ReturnLogsEditController = require('../controllers/return-logs-edit.controller.js')
-
-const basePath = '/return-log-edit/'
-
-const routes = [
-  {
-    method: 'GET',
-    path: basePath + 'setup',
-    options: {
-      handler: ReturnLogsEditController.setup,
-      auth: {
-        access: {
-          scope: ['billing']
-        }
-      }
-    }
-  }
-]
-
-module.exports = routes
